move pages to corresponding folder
