+++ src/components/shared/Footer/CompanyInfo/CompanyInfo.styles.scss
@@ -141,7 +141,8 @@
     .offices-info {
       grid-row: 2 / 2;
       margin-bottom: 0px;
-      padding-right: 20px;
+      width: 90%;
+      margin-right: 10%;
 
       .city {
         font-size: 32px;
